float16 broke the anomaly z-score: every dynamic channel would train on zeros
Found while fixing E-038's memory diagnosis, and it is the more dangerous of
the two: #365 died on host RAM BEFORE reaching this, so fixing only the memory
would have unmasked it into a run that looked healthy and meant nothing.

numpy upcasts the accumulator for np.mean on float16 but NOT for np.std/np.var
-- _methods._var upcasts only integer and bool. trainprobe.anomaly_transform
z-scores each dynamic channel by v.std() over the train pool: ~204M entries on
family 4. The sum of squared residuals passes 65504, v.std() returns inf, and

    X[..., c] = (X[..., c] - v.mean()) / (v.std() + 1e-6)

is exactly 0.0 everywhere. Measured 2026-08-17: a float16 pool of 1M already
overflows; in float64 the same pool reads 0.999844.

Family 3 is float32 (build_family3.py: open_memmap dtype=np.float32, 516 x 281
x 481 x 39 = 10.9 GB) and never reached the limit, so six weeks of runs were
correct. Family 4 is the FIRST float16 tensor -- build_family4.py chose it so
3142 x 281 x 481 x 39 fits in 33.1 GB rather than 66.3. The dtype decision
that made the tensor buildable is what broke the transform, and the workflow
defaults anomaly to "true", so the very next dispatch would have hit it.

Nothing downstream would have said so. The loss is finite, gpu_util is normal,
the light probe returns a number. Worse than a wasted dispatch: probe_kfold.py
calls this same function, so E-038's frozen control -- "the cheapest number in
the plan and the only one that can falsify the premise" -- would have scored
the monthly codec against an all-zero tensor and returned a null, which reads
exactly like the premise being falsified. All three arms, one silent cause.

Fixed by naming dtype=float64 on every reduction in anomaly_transform: the
dynamic-channel detection (np.nanmean/np.nanstd, which overflow the same way
on a channel with a nonzero spatial mean), the climatology, and the z-score.
The float32 path moves by 5.2e-9 relative -- seven orders below the sd 0.123
seed noise ml/CLAUDE.md section 3 records for the stage-2 probe.

train.py carried a hand-inlined SECOND copy of the whole transform, so the bug
existed in two places and would have had to be found twice. The duplicate is
deleted; train.py now calls trainprobe.anomaly_transform, which is also what
makes the frozen control comparable to the trained arms at all.

tests/test_e038_anomaly_dtype.py pins it. Note the fixture had to be MEASURED
into regime, not assumed: the first version (T=600, 0.5 noise) had a 180k-entry
pool summing to ~45k squared residuals -- under the limit -- and passed against
the unfixed code, pinning nothing. _assert_in_regime now runs numpy's own
default-accumulator std over the real pool and fails if it comes back finite.
Verified by reverting the fix: 3 float16 tests fail, 3 float32 tests pass.
Full python suite 87 passed.

+++ ml/trainprobe.py
@@ -46,13 +46,30 @@ def anomaly_transform(X, moy, t_hold, x_hold):
     channels become departures from their own train-years monthly
     climatology, then z-scored on train data. Returns (X, dynamic)."""
     T, H, W, C = X.shape
+    # EVERY reduction below names dtype=float64, and that is not cosmetic.
+    # Family 4 is the project's first float16 tensor (build_family4.py chose
+    # it to fit 33.1 GB rather than 66.3). numpy upcasts the accumulator for
+    # np.mean on float16 but NOT for np.std/np.var — _methods._var only
+    # upcasts integer and bool. The z-score at the end of this function sums
+    # ~204M squared residuals; in float16 that passes 65504, returns inf, and
+    # (X - mu) / (inf + 1e-6) is exactly 0.0. Every dynamic channel would
+    # become zeros while every loss, gpu_util and probe still looked healthy —
+    # and probe_kfold.py calls this function, so E-038's frozen control would
+    # have "falsified" its premise against an all-zero tensor. Family 3 was
+    # float32 and never reached the limit. Measured 2026-08-17: float16 pool
+    # of 30M -> inf; the same pool in float64 -> 0.999844. The float32 path
+    # moves by 5.2e-9 relative, ~7 orders below the sd 0.123 seed noise the
+    # stage-2 probe already carries (ml/CLAUDE.md §3).
     dynamic = [c for c in range(C)
-               if np.nanstd(np.nanmean(X[..., c], axis=(1, 2))) > 1e-6]
+               if np.nanstd(np.nanmean(X[..., c], axis=(1, 2),
+                                       dtype=np.float64),
+                            dtype=np.float64) > 1e-6]
     clim = np.full((12, H, W, C), np.nan, dtype=np.float32)
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
         for m in range(12):
-            clim[m] = np.nanmean(X[(moy == m) & ~t_hold], axis=0)
+            clim[m] = np.nanmean(X[(moy == m) & ~t_hold], axis=0,
+                                 dtype=np.float64)
     for c in dynamic:
         # clim[moy, :, :, c] is [T,H,W]; the equivalent-looking
         # clim[moy][..., c] materialises the whole [T,H,W,C] fancy index and
@@ -61,7 +78,8 @@ def anomaly_transform(X, moy, t_hold, x_hold):
         X[..., c] = X[..., c] - clim[moy, :, :, c]
         v = X[..., c][np.isfinite(X[..., c]) & ~t_hold[:, None, None]
                       & ~x_hold[None, None, :]]
-        X[..., c] = (X[..., c] - v.mean()) / (v.std() + 1e-6)
+        X[..., c] = (X[..., c] - v.mean(dtype=np.float64)) / (
+            v.std(dtype=np.float64) + 1e-6)
     return X, dynamic
